\name{fd.get.available.methods}
\alias{fd.get.available.methods}
\title{Available Estimation Methods}
\description{The function returns a list of estimation methods that can be used in \code{\link{fd.estimate}} and other functions of the package.}
\usage{fd.get.available.methods(dim = 1)}
\arguments{
\item{dim}{Dimension of data for which the estimation methods should be obtained.}	
}
\value{A list of the available methods. Their order number in the list corresponds to the internal codes of the methods.}  
\seealso{\code{\link{fd.estimate}}, \code{\link{fd.get}}}
\keyword{programming}

